build docs automatically via Travis CI
This uses the approach from
https://gist.github.com/domenic/ec8b0fc8ab45f39403dd where a custom
SSH key is used for deploying. The SSH private key is stored in
encrypted form in the repo itself, but it can only be decrypted with
another key that is kept secret by Travis CI. The generated docs end
up in the gh-pages branch of the repo.

The generated docs get removed. This will make merging PRs a lot
simpler (no more rebasing because of conflicts in the generated HTML).

Fixes: #79